init wfs pkg refactor
