+++ pkg/wfs200/exception/exception_test.go
@@ -1,4 +1,4 @@
-package wfs200
+package exception
 
 import (
 	"testing"
