+++ pkg/wfs200/request/describefeaturetype.go
@@ -1,4 +1,4 @@
-package wfs200
+package request
 
 import (
 	"encoding/xml"
